+++ src/ViewUser.js
@@ -16,6 +16,7 @@ import {
   Col,
   IfPermission,
   IfInterface,
+  TextField,
   Layer,
   Headline,
   AccordionSet,
@@ -30,6 +31,7 @@ import LoanActionsHistory from './LoanActionsHistory';
 import { ChargeFeeFine } from './components/Accounts';
 import AccountsHistory from './AccountsHistory';
 import AccountActionsHistory from './AccountActionsHistory';
+import PatronBlockLayer from './components/PatronBlock';
 
 import { toListAddresses, toUserAddresses } from './converters/address';
 import { getFullName, eachPromise } from './util';
@@ -42,6 +44,7 @@ import {
   ExtendedInfo,
   ContactInfo,
   ProxyPermissions,
+  PatronBlock,
   UserPermissions,
   UserLoans,
   UserRequests,
@@ -60,6 +63,11 @@ class ViewUser extends React.Component {
         const { path } = action.meta;
         return refresh || path.match(/link/);
       },
+    },
+    hasPatronBlocks: {
+      type: 'okapi',
+      records: 'manualblocks',
+      path: 'manualblocks?query=(userId=:{id})&limit=1',
     },
     loansHistory: {
       type: 'okapi',
@@ -173,6 +181,7 @@ class ViewUser extends React.Component {
       viewOpenLoansMode: true,
       selectedLoan: {},
       selectedAccount: {},
+      selectedPatronBlock: {},
       addRecord: false,
       lastUpdate: null,
       sections: {
@@ -180,6 +189,7 @@ class ViewUser extends React.Component {
         extendedInfoSection: false,
         contactInfoSection: false,
         proxySection: false,
+        patronBlocksSection: false,
         loansSection: false,
         requestsSection: false,
         accountsSection: false,
@@ -191,6 +201,7 @@ class ViewUser extends React.Component {
     this.connectedUserLoans = props.stripes.connect(UserLoans);
     this.connectedUserRequests = props.stripes.connect(UserRequests);
     this.connectedUserAccounts = props.stripes.connect(UserAccounts);
+    this.connectedPatronBlock = props.stripes.connect(PatronBlock);
     this.connectedLoansHistory = props.stripes.connect(LoansHistory);
     this.connectedLoanActionsHistory = props.stripes.connect(LoanActionsHistory);
     this.connectedUserInfo = props.stripes.connect(UserInfo);
@@ -220,6 +231,9 @@ class ViewUser extends React.Component {
     this.onClickCloseAccountsHistory = this.onClickCloseAccountsHistory.bind(this);
     this.onClickViewAccountActionsHistory = this.onClickViewAccountActionsHistory.bind(this);
     this.onClickCloseAccountActionsHistory = this.onClickCloseAccountActionsHistory.bind(this);
+    this.connectedPatronBlockLayer = props.stripes.connect(PatronBlockLayer);
+    this.onClickViewPatronBlock = this.onClickViewPatronBlock.bind(this);
+    this.onClickClosePatronBlock = this.onClickClosePatronBlock.bind(this);
     this.showCallout = null;
 
     this.editButton = React.createRef();
@@ -388,6 +402,27 @@ class ViewUser extends React.Component {
     });
   }
 
+  onClickViewPatronBlock(e, mode, selectedPatronBlock) {
+    if (e) e.preventDefault();
+    const layer = (mode === 'add') ? 'add-block' : 'edit-block';
+    if (mode === 'add') {
+      this.props.mutator.query.update({ layer });
+    } else {
+      this.setState({
+        selectedPatronBlock,
+      });
+      this.props.mutator.query.update({ layer, block: selectedPatronBlock.id });
+    }
+  }
+
+  onClickClosePatronBlock(e) {
+    if (e) e.preventDefault();
+    this.props.mutator.query.update({ layer: null, block: null });
+    this.setState({
+      selectedPatronBlock: {},
+    });
+  }
+
   handleAddRecords() {
     const add = this.state.addRecord;
     this.setState({
@@ -398,7 +433,8 @@ class ViewUser extends React.Component {
   getUser() {
     const { resources, match: { params: { id } } } = this.props;
     const selUser = (resources.selUser || {}).records || [];
-    if (!selUser || selUser.length === 0 || !id) return null;
+
+    if (selUser.length === 0 || !id) return null;
     // Logging below shows this DOES sometimes find the wrong record. But why?
     // console.log(`getUser: found ${selUser.length} users, id '${selUser[0].id}' ${selUser[0].id === id ? '==' : '!='} '${id}'`);
     return selUser.find(u => u.id === id);
@@ -533,16 +569,24 @@ class ViewUser extends React.Component {
     }
   }
 
+  isLayerOpen = value => {
+    const { layer } = this.props.resources.query;
+
+    return layer === value;
+  };
+
   render() {
     const {
       resources,
       stripes,
       parentResources,
       tagsEnabled,
+      location
     } = this.props;
 
     const addressTypes = (parentResources.addressTypes || {}).records || [];
-    const query = resources.query;
+    const query = queryString.parse(location.search || '');
+
     const user = this.getUser();
     const tags = ((user && user.tags) || {}).tagList || [];
     const patronGroups = (resources.patronGroups || {}).records || [];
@@ -559,7 +603,7 @@ class ViewUser extends React.Component {
           {
             tagsEnabled &&
             <FormattedMessage id="ui-users.showTags">
-              { ariaLabel => (
+              {ariaLabel => (
                 <IconButton
                   icon="tag"
                   id="clickable-show-tags"
@@ -572,7 +616,7 @@ class ViewUser extends React.Component {
           }
           <IfPermission perm="users.item.put">
             <FormattedMessage id="ui-users.crud.editUser">
-              { ariaLabel => (
+              {ariaLabel => (
                 <IconButton
                   icon="edit"
                   id="clickable-edituser"
@@ -603,29 +647,32 @@ class ViewUser extends React.Component {
       );
     }
 
+    const hasPatronBlocks = (get(resources, ['hasPatronBlocks', 'isPending'], true)) ? -1 : 1;
+    const totalPatronBlocks = get(resources, ['hasPatronBlocks', 'other', 'totalRecords'], 0);
     const patronGroupId = get(user, ['patronGroup'], '');
     const patronGroup = patronGroups.find(g => g.id === patronGroupId) || { group: '' };
     const addresses = toListAddresses(get(user, ['personal', 'addresses'], []), addressTypes);
     const userFormData = this.getUserFormData(user, addresses, sponsors, proxies, permissions, servicePoints, preferredServicePoint);
 
-    const loansHistory = (<this.connectedLoansHistory
-      buildRecords={this.buildRecords}
-      user={user}
-      loansHistory={loans}
-      patronGroup={patronGroup}
-      stripes={stripes}
-      history={this.props.history}
-      onCancel={this.onClickCloseLoansHistory}
-      onClickViewOpenLoans={this.onClickViewOpenLoans}
-      onClickViewClosedLoans={this.onClickViewClosedLoans}
-      onClickViewLoanActionsHistory={this.onClickViewLoanActionsHistory}
-      onClickViewChargeFeeFine={this.onClickViewChargeFeeFine}
-      onClickViewOpenAccounts={this.onClickViewOpenAccounts}
-      onClickViewAccountActionsHistory={this.onClickViewAccountActionsHistory}
-      onClickViewClosedAccounts={this.onClickViewClosedAccounts}
-      onClickViewAllAccounts={this.onClickViewAllAccounts}
-      openLoans={query.layer === 'open-loans'}
-    />);
+    const loansHistory = (
+      <this.connectedLoansHistory
+        buildRecords={this.buildRecords}
+        user={user}
+        loansHistory={loans}
+        patronGroup={patronGroup}
+        stripes={stripes}
+        history={this.props.history}
+        onCancel={this.onClickCloseLoansHistory}
+        onClickViewOpenLoans={this.onClickViewOpenLoans}
+        onClickViewClosedLoans={this.onClickViewClosedLoans}
+        onClickViewLoanActionsHistory={this.onClickViewLoanActionsHistory}
+        onClickViewChargeFeeFine={this.onClickViewChargeFeeFine}
+        onClickViewOpenAccounts={this.onClickViewOpenAccounts}
+        onClickViewAccountActionsHistory={this.onClickViewAccountActionsHistory}
+        onClickViewClosedAccounts={this.onClickViewClosedAccounts}
+        onClickViewAllAccounts={this.onClickViewAllAccounts}
+        openLoans={query.layer === 'open-loans'}
+      />);
 
     const loanDetails = (
       <this.connectedLoanActionsHistory
@@ -663,8 +710,12 @@ class ViewUser extends React.Component {
 
           <Headline size="xx-large" tag="h2">{getFullName(user)}</Headline>
 
-          <Row end="xs">
-            <Col xs>
+          <Row>
+            <Col xs={10}>
+              {(hasPatronBlocks === 1 && totalPatronBlocks > 0) ?
+                <TextField fullWidth readOnly value="Patron has blocks in place" error=" " /> : ''}
+            </Col>
+            <Col xs={2}>
               <ExpandAllButton accordionStatus={this.state.sections} onToggle={this.handleExpandAll} />
             </Col>
           </Row>
@@ -677,6 +728,16 @@ class ViewUser extends React.Component {
               stripes={stripes}
               expanded={this.state.sections.userInformationSection}
               onToggle={this.handleSectionToggle}
+            />
+            <this.connectedPatronBlock
+              accordionId="patronBlocksSection"
+              hasPatronBlocks={(hasPatronBlocks === 1 && totalPatronBlocks > 0)}
+              expanded={this.state.sections.patronBlocksSection}
+              onToggle={this.handleSectionToggle}
+              onClickViewPatronBlock={this.onClickViewPatronBlock}
+              addRecord={this.state.addRecord}
+              initialValues={userFormData}
+              {...this.props}
             />
             <ExtendedInfo
               accordionId="extendedInfoSection"
@@ -773,9 +834,9 @@ class ViewUser extends React.Component {
             </IfPermission>
           </AccordionSet>
           <FormattedMessage id="ui-users.editUserDialog">
-            { contentLabel => (
+            {contentLabel => (
               <Layer
-                isOpen={query.layer ? query.layer === 'edit' : false}
+                isOpen={this.isLayerOpen('edit')}
                 contentLabel={contentLabel}
                 afterClose={this.afterCloseEdit}
               >
@@ -793,8 +854,22 @@ class ViewUser extends React.Component {
               </Layer>
             )}
           </FormattedMessage>
+          <Layer isOpen={query.layer ? query.layer === 'add-block' || query.layer === 'edit-block' : false} label={query.layer === 'add-block' ? 'Add Block' : 'Edit Block'}>
+            <this.connectedPatronBlockLayer
+              {...this.props}
+              query={query}
+              user={user}
+              selectedPatronBlock={this.state.selectedPatronBlock}
+              handleAddRecords={this.handleAddRecords}
+              onCancel={this.onClickClosePatronBlock}
+            />
+          </Layer>
           <Layer
-            isOpen={query.layer ? query.layer === 'open-accounts' || query.layer === 'closed-accounts' || query.layer === 'all-accounts' : false}
+            isOpen={
+              this.isLayerOpen('open-accounts') ||
+              this.isLayerOpen('closed-accounts') ||
+              this.isLayerOpen('all-accounts')
+            }
             label={<FormattedMessage id="ui-users.accounts.title" />}
           >
             <this.connectedAccountsHistory
@@ -815,7 +890,10 @@ class ViewUser extends React.Component {
               onClickCloseAccountActionsHistory={this.onClickCloseAccountActionsHistory}
             />
           </Layer>
-          <Layer isOpen={query.layer ? query.layer === 'charge' : false} label="Charge Fee/Fine">
+          <Layer
+            isOpen={this.isLayerOpen('charge')}
+            label={<FormattedMessage id="ui-users.chargeFeefine" />}
+          >
             <this.connectedCharge
               servicePoints={servicePoints}
               preferredServicePoint={preferredServicePoint}
@@ -827,7 +905,10 @@ class ViewUser extends React.Component {
               handleAddRecords={this.handleAddRecords}
             />
           </Layer>
-          <Layer isOpen={query.layer ? query.layer === 'account' : false} label="Account Actions History">
+          <Layer
+            isOpen={this.isLayerOpen('account')}
+            label={<FormattedMessage id="ui-users.accountActionHistory" />}
+          >
             <this.connectedAccountActionsHistory
               user={user}
               patronGroup={patronGroup}
@@ -846,9 +927,9 @@ class ViewUser extends React.Component {
 
           <IfPermission perm="ui-users.loans.all">
             <FormattedMessage id="ui-users.loans.title">
-              { contentLabel => (
+              {contentLabel => (
                 <Layer
-                  isOpen={query.layer ? query.layer === 'open-loans' || query.layer === 'closed-loans' : false}
+                  isOpen={this.isLayerOpen('open-loans') || this.isLayerOpen('closed-loans')}
                   contentLabel={contentLabel}
                 >
                   {loansHistory}
@@ -857,9 +938,9 @@ class ViewUser extends React.Component {
             </FormattedMessage>
 
             <FormattedMessage id="ui-users.loanActionsHistory">
-              { contentLabel => (
+              {contentLabel => (
                 <Layer
-                  isOpen={query.layer ? query.layer === 'loan' : false}
+                  isOpen={this.isLayerOpen('loan')}
                   contentLabel={contentLabel}
                 >
                   {loanDetails}