+++ src/components/Accounts/AllAccounts/AllAccounts.js
@@ -3,6 +3,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import {
   FormattedMessage,
+  intlShape,
   FormattedTime,
   FormattedDate,
 } from 'react-intl';
@@ -33,6 +34,7 @@ class AllAccounts extends React.Component {
     onClickViewLoanActionsHistory: PropTypes.func.isRequired,
     accounts: PropTypes.arrayOf(PropTypes.object),
     onChangeActions: PropTypes.func.isRequired,
+    intl: intlShape.isRequired,
   };
 
   constructor(props) {
@@ -46,36 +48,26 @@ class AllAccounts extends React.Component {
     this.getLoan = this.getLoan.bind(this);
 
     this.sortMap = {
-      'created': f => (f.metadata || {}).createdDate,
-      'updated': f => (f.metadata || {}).updatedDate,
-      'type': f => f.feeFineType,
+      'metadata.createdDate': f => (f.metadata || {}).createdDate,
+      'metadata.updatedDate': f => (f.metadata || {}).updatedDate,
+      'feeFineType': f => f.feeFineType,
       'amount': f => f.amount,
       'remaining': f => f.remaining,
-      'status': f => (f.paymentStatus || {}).name,
-      'owner': f => f.feeFineOwner,
+      'paymentStatus.name': f => (f.paymentStatus || {}).name,
+      'feeFineOwner': f => f.feeFineOwner,
       'title': f => f.title,
       'barcode': f => f.barcode,
-      'number': f => f.callNumber,
-      'due': f => f.dueDate,
-      'returned': f => f.returnedDate,
+      'callNumber': f => f.callNumber,
+      'dueDate': f => f.dueDate,
+      'returnedDate': f => f.returnedDate,
     };
 
     this.state = {
       checkedAccounts: {},
       allChecked: false,
       sortOrder: [
-        'created',
-        'updated',
-        'type',
-        'amount',
-        'remaining',
-        'status',
-        'owner',
-        'title',
-        'barcode',
-        'number',
-        'due',
-        'returned',
+        'metadata.createdDate',
+        'metadata.createdDate',
       ],
       sortDirection: ['desc', 'desc'],
     };
@@ -104,12 +96,12 @@ class AllAccounts extends React.Component {
   }
 
   onSort(e, meta) {
-    if (!this.sortMap[meta.alias]) return;
+    if (!this.sortMap[meta.name]) return;
 
     let { sortOrder, sortDirection } = this.state;
 
-    if (sortOrder[0] !== meta.alias) {
-      sortOrder = [meta.alias, sortOrder[1]];
+    if (sortOrder[0] !== meta.name) {
+      sortOrder = [meta.name, sortOrder[1]];
       sortDirection = ['asc', sortDirection[1]];
     } else {
       const direction = (sortDirection[0] === 'desc') ? 'asc' : 'desc';
@@ -184,18 +176,18 @@ class AllAccounts extends React.Component {
           type="checkbox"
         />
       ),
-      'created': f => (f.metadata ? <FormattedDate value={f.metadata.createdDate} /> : '-'),
-      'updated': f => (f.metadata && f.metadata.createdDate !== f.metadata.updatedDate ? <FormattedDate value={f.metadata.updatedDate} /> : '-'),
-      'type': f => (f.feeFineType ? this.comments(f) : '-'),
+      'metadata.createdDate': f => (f.metadata ? <FormattedDate value={f.metadata.createdDate} /> : '-'),
+      'metadata.updatedDate': f => (f.metadata && f.metadata.createdDate !== f.metadata.updatedDate ? <FormattedDate value={f.metadata.updatedDate} /> : '-'),
+      'feeFineType': f => (f.feeFineType ? this.comments(f) : '-'),
       'amount': f => (f.amount ? parseFloat(f.amount).toFixed(2) : '-'),
       'remaining': f => parseFloat(f.remaining).toFixed(2) || '0.00',
-      'status': f => (f.paymentStatus || {}).name || '-',
-      'owner': f => (f.feeFineOwner ? f.feeFineOwner : '-'),
+      'paymentStatus.name': f => (f.paymentStatus || {}).name || '-',
+      'feeFineOwner': f => (f.feeFineOwner ? f.feeFineOwner : '-'),
       'title': f => (f.title ? `${f.title} (${f.materialType})` : '-'),
       'barcode': f => (f.barcode ? f.barcode : '-'),
-      'number': f => (f.callNumber ? f.callNumber : '-'),
-      'due': f => (f.dueDate ? this.formatDateTime(f.dueDate) : '-'),
-      'returned': f => (f.returnedDate ? this.formatDateTime(f.returnedDate) : this.formatDateTime(this.getLoan(f).returnDate) || '-'),
+      'callNumber': f => (f.callNumber ? f.callNumber : '-'),
+      'dueDate': f => (f.dueDate ? this.formatDateTime(f.dueDate) : '-'),
+      'returnedDate': f => (f.returnedDate ? this.formatDateTime(f.returnedDate) : this.formatDateTime(this.getLoan(f).returnDate) || '-'),
       ' ': f => this.renderActions(f),
     };
   }
@@ -374,8 +366,22 @@ class AllAccounts extends React.Component {
     const props = this.props;
 
     const fees = _.orderBy(props.accounts, [this.sortMap[sortOrder[0]], this.sortMap[sortOrder[1]]], sortDirection);
+    const { intl } = this.props;
+
     const columnMapping = {
       '  ': (<input type="checkbox" checked={allChecked} name="check-all" onChange={this.toggleAll} />),
+      'metadata.createdDate': intl.formatMessage({ id: 'ui-users.accounts.history.columns.created' }),
+      'metadata.updatedDate': intl.formatMessage({ id: 'ui-users.accounts.history.columns.updated' }),
+      'feeFineType': intl.formatMessage({ id: 'ui-users.accounts.history.columns.type' }),
+      'amount': intl.formatMessage({ id: 'ui-users.accounts.history.columns.amount' }),
+      'remaining': intl.formatMessage({ id: 'ui-users.accounts.history.columns.remaining' }),
+      'paymentStatus.name': intl.formatMessage({ id: 'ui-users.accounts.history.columns.status' }),
+      'feeFineOwner': intl.formatMessage({ id: 'ui-users.accounts.history.columns.owner' }),
+      'title': intl.formatMessage({ id: 'ui-users.accounts.history.columns.title' }),
+      'barcode': intl.formatMessage({ id: 'ui-users.accounts.history.columns.barcode' }),
+      'callNumber': intl.formatMessage({ id: 'ui-users.accounts.history.columns.number' }),
+      'dueDate': intl.formatMessage({ id: 'ui-users.accounts.history.columns.due' }),
+      'returnedDate': intl.formatMessage({ id: 'ui-users.accounts.history.columns.returned' }),
     };
 
     return (
@@ -386,12 +392,12 @@ class AllAccounts extends React.Component {
           columnMapping={columnMapping}
           columnWidths={{
             '  ': 28,
-            'created': 110,
-            'type': 200,
-            'updated': 110,
+            'metadata.createdDate': 110,
+            'feeFineType': 180,
+            'metadata.updatedDate': 110,
             'barcode': 120,
-            'due': 110,
-            'returned': 110
+            'dueDate': 110,
+            'returnedDate': 110
           }}
           visibleColumns={this.props.visibleColumns}
           fullWidth