+++ src/settings/OwnerSettings.js
@@ -10,6 +10,10 @@ import {
   ConfirmationModal,
   Pane,
   Paneset,
+  Modal,
+  Row,
+  Button,
+  Col,
   MultiSelection,
 } from '@folio/stripes/components';
 import { Field } from 'redux-form';
@@ -32,7 +36,7 @@ class OwnerSettings extends React.Component {
         path: 'owners/%{activeRecord.id}',
       },
       GET: {
-        path: 'owners?query=cql.allRecords=1 &limit=500'
+        path: 'owners?query=cql.allRecords=1 sortby owner&limit=500'
       }
     },
     servicePoints: {
@@ -52,6 +56,7 @@ class OwnerSettings extends React.Component {
 
     this.state = {
       showConfirmDialog: false,
+      showItemInUseDialog: false,
       selectedItem: {},
       primaryField: 'owner',
     };
@@ -80,6 +85,7 @@ class OwnerSettings extends React.Component {
         this.deleteItemResolve();
       })
       .catch(() => {
+        this.setState({ showItemInUseDialog: true });
         this.deleteItemReject();
       })
       .finally(() => this.hideConfirmDialog());
@@ -99,6 +105,7 @@ class OwnerSettings extends React.Component {
 
   hideItemInUseDialog() {
     this.setState({
+      showItemInUseDialog: false,
       selectedItem: {},
     });
   }
@@ -134,7 +141,7 @@ class OwnerSettings extends React.Component {
   validate({ items }) {
     const { primaryField } = this.state;
     const servicePoints = _.get(this.props.resources, ['servicePoints', 'records', 0, 'servicepoints'], []);
-    const none = servicePoints.find(s => s.name === 'Circ Desk 1') || {};
+    const none = servicePoints.find(s => s.name === 'None') || {};
 
     if (Array.isArray(items)) {
       const errors = [];
@@ -151,7 +158,7 @@ class OwnerSettings extends React.Component {
         const asp = item.servicePointOwner || [];
         asp.forEach(s => {
           if (s.value === none.id && asp.length > 1) {
-            itemErrors.servicePointOwner = 'Error';
+            itemErrors.servicePointOwner = <FormattedMessage id="ui-users.owners.error" />;
           }
         });
 
@@ -183,7 +190,7 @@ class OwnerSettings extends React.Component {
 
         asp.forEach(s => {
           if (s.value === none.id) {
-            itemWarning.servicePointOwner = 'Warning: Overdue fines/lost item fees will not be collected without a service point';
+            itemWarning.servicePointOwner = <FormattedMessage id="ui-users.owners.warning" />;
           }
         });
 
@@ -197,7 +204,8 @@ class OwnerSettings extends React.Component {
 
   render() {
     if (!this.props.resources.owners) return <div />;
-    const { intl } = this.props;
+    const { intl: { formatMessage } } = this.props;
+    const type = formatMessage({ id: 'ui-users.owners.singular' });
     const rows = this.props.resources.owners.records || [];
     const term = this.state.selectedItem[this.state.primaryField];
     const servicePoints = _.get(this.props.resources, ['servicePoints', 'records', 0, 'servicepoints'], []);
@@ -233,7 +241,7 @@ class OwnerSettings extends React.Component {
       <SafeHTMLMessage
         id="stripes-smart-components.cv.termWillBeDeleted"
         values={{
-          type: intl.formatMessage({ id: 'ui-users.owners.singular' }),
+          type: formatMessage({ id: 'ui-users.owners.singular' }),
           term,
         }}
       />
@@ -248,17 +256,25 @@ class OwnerSettings extends React.Component {
     };
     return (
       <Paneset>
-        <Pane defaultWidth="fill" fluidContentWidth paneTitle={this.props.label}>
+        <Pane
+          defaultWidth="fill"
+          fluidContentWidth
+          height="100%"
+          paneTitle={formatMessage({ id: 'ui-users.owners.label' })}
+        >
           <EditableList
             {...this.props}
+            label={formatMessage({ id: 'ui-users.owners.label' })}
+            height="600"
+            virtualize
             fieldComponents={fieldComponents}
             contentData={rows}
             createButtonLabel={<FormattedMessage id="stripes-core.button.new" />}
             visibleFields={['owner', 'desc', 'servicePointOwner']}
             columnMapping={{
-              'owner': intl.formatMessage({ id: 'ui-users.owners.columns.owner' }),
-              'desc': intl.formatMessage({ id: 'ui-users.owners.columns.desc' }),
-              'servicePointOwner': intl.formatMessage({ id: 'ui-users.owners.columns.asp' }),
+              'owner': formatMessage({ id: 'ui-users.owners.columns.owner' }),
+              'desc': formatMessage({ id: 'ui-users.owners.columns.desc' }),
+              'servicePointOwner': formatMessage({ id: 'ui-users.owners.columns.asp' }),
             }}
             formatter={formatter}
             onUpdate={this.onUpdateItem}
@@ -283,7 +299,7 @@ class OwnerSettings extends React.Component {
             heading={(
               <FormattedMessage
                 id="stripes-core.button.deleteEntry"
-                values={{ entry: intl.formatMessage({ id: 'ui-users.owners.singular' }) }}
+                values={{ entry: formatMessage({ id: 'ui-users.owners.singular' }) }}
               />
             )}
             message={modalMessage}
@@ -291,6 +307,24 @@ class OwnerSettings extends React.Component {
             onCancel={this.hideConfirmDialog}
             confirmLabel={<FormattedMessage id="stripes-core.button.delete" />}
           />
+          <Modal
+            open={this.state.showItemInUseDialog}
+            label={<FormattedMessage id="stripes-smart-components.cv.cannotDeleteTermHeader" values={{ type }} />}
+            size="small"
+          >
+            <Row>
+              <Col xs>
+                <FormattedMessage id="stripes-smart-components.cv.cannotDeleteTermMessage" values={{ type }} />
+              </Col>
+            </Row>
+            <Row>
+              <Col xs>
+                <Button buttonStyle="primary" onClick={this.hideItemInUseDialog}>
+                  <FormattedMessage id="stripes-core.label.okay" />
+                </Button>
+              </Col>
+            </Row>
+          </Modal>
           <Callout ref={(ref) => { this.callout = ref; }} />
         </Pane>
       </Paneset>