+++ CHANGELOG.md
@@ -30,6 +30,14 @@
 * Link to a user's requests. Fixes UIU-677.
 * Bring back handleSubmit when using submit button. Fixes UIU-743.
 * Fix permission menu after reopening. Fixes UIU-739.
+* Apply internationalization to all hardcoded strings for Settings. Fixes UIU-691.
+* Pass `userBarcode` to new requests. Fixes UIU-690. Refs UIREQ-146.
+* Optimize code for Manual Fees/Fines Table. Fix UIU-198.
+* Modify associated service points for . Ref UIU-611.
+* Create permission "Settings (Users): Can create, edit and remove fee/fine settings". Ref UIU-584.
+* Create permission "Fees/Fines: All Actions permissions". Ref UIU-673.
+* Implement CRUD Manual Patron Block. Ref UIU-674.
+* Create permission "Patron blocks: All permissions". Ref UIU-728.
 * Provide `renderToOverlay` to `<MultiSelection>` to avoid clipping. Fixes STSMACOM-149.
 
 ## [2.17.0](https://github.com/folio-org/ui-users/tree/v2.17.0) (2018-10-5)